Create SentimentAnalysis_on_TwitterData.R
I have made this sentiment analysis on the data of Twitter. As we all know that Twitter is a social media platform that enables its users to be expressive with their thoughts, so doing sentiment analysis on the data of Twitter was quite interesting. 

In this project, I have used the R studio for coding purposes and Kaggle.com to collect the dataset.